requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Mail==0.9.1
python-dotenv==1.0.0
Werkzeug==2.3.6
itsdangerous==2.1.2
qrcode==7.4.2
pandas==2.0.3
WeasyPrint==58.0
gunicorn==20.1.0